Signedness error

+++ src/stubby.c
@@ -256,7 +256,7 @@ static getdns_return_t parse_config_file(const char *fn)
 	}
 	rewind(fh);
 	read_sz = fread(config_file, 1, config_file_sz + 1, fh);
-	if (read_sz > config_file_sz || ferror(fh) || !feof(fh)) {
+	if (read_sz > (size_t)config_file_sz || ferror(fh) || !feof(fh)) {
 		fprintf( stderr, "An error occurred while reading \"%s\": %s\n"
 		       , fn, strerror(errno));
 		fclose(fh);
